+++ src/main/kotlin/com/christopherosthues/starwarsprogressbar/constants/PluginConstants.kt
@@ -2,6 +2,8 @@ package com.christopherosthues.starwarsprogressbar.constants
 
 internal object PluginConstants {
     const val PluginId = "com.christopherosthues.starwarsprogressbar"
+    private const val GotItTooltipIdPrefix = "com.christopherosthues.starwarsprogressbar.gotit."
+    const val GotItIconSelectionId = GotItTooltipIdPrefix + "iconselection"
     const val PluginSearchId = "com.christopherosthues.starwarsprogressbar.ui.configuration"
     const val NotificationGroupId = "Star Wars Progress Bar updated"
 }