+++ src/main/kotlin/com/christopherosthues/starwarsprogressbar/notification/GotItService.kt
@@ -1,0 +1,13 @@
+package com.christopherosthues.starwarsprogressbar.notification
+
+object GotItService {
+    private val gotItTooltips: MutableList<GotIt> = mutableListOf()
+
+    fun addGotItMessage(tooltip: GotIt) {
+        gotItTooltips.add(tooltip)
+    }
+
+    fun showGotItTooltips() {
+        gotItTooltips.forEach { it.show() }
+    }
+}