+++ src/main/kotlin/com/christopherosthues/starwarsprogressbar/ui/components/VehiclesPanel.kt
@@ -2,10 +2,13 @@ package com.christopherosthues.starwarsprogressbar.ui.components
 
 import com.christopherosthues.starwarsprogressbar.constants.BundleConstants
 import com.christopherosthues.starwarsprogressbar.StarWarsBundle
+import com.christopherosthues.starwarsprogressbar.constants.PluginConstants
 import com.christopherosthues.starwarsprogressbar.models.StarWarsFaction
 import com.christopherosthues.starwarsprogressbar.models.FactionHolder
 import com.christopherosthues.starwarsprogressbar.ui.configuration.StarWarsState
 import com.christopherosthues.starwarsprogressbar.ui.events.VehicleClickListener
+import com.intellij.ui.GotItTooltip
+import com.intellij.ui.roots.ScalableIconComponent
 import com.intellij.util.ui.JBUI
 import com.intellij.util.ui.ThreeStateCheckBox
 import java.awt.BorderLayout