+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Ahsoka Tano's and Anakin Skywalker's Delta-7b to Jedi Order faction
 - Ahsoka Tano's and Obi-Wan Kenobi's Eta-2 Actis to Jedi Order faction
 - Update preview when the icon of a vehicle was clicked
+- Got it tooltip to settings for new features
 
 ## [1.2.0] - 2022-07-18
 ### Changed